pyproject.toml
# SPDX-License-Identifier: LGPL-3.0-or-later

[build-system]
requires = [
    "setuptools>=45",
    "wheel",
    "setuptools_scm[toml]>=6.2",
    "hatch-fancy-pypi-readme",
]
build-backend = "setuptools.build_meta"

[project]
name = "torch_dplr"
dynamic = ["version", "readme"]
description = "DPLR in PyTorch backend"
license = { file = ".license-header.txt" }
dependencies = [
    "deepmd-kit>=3.1.2",
    "torch>=2.1.0",
    "torchopt==0.7.3",
    "numpy>=1.20.0, <2.0",
]
[project.urls]
Homepage = "https://github.com/ChiahsinChu/torch-dplr"

[project.optional-dependencies]
torch_admp = [
    "torch_admp @ git+https://github.com/chiahsinchu/torch-admp.git@v1.1.0a",
]

test = ["pytest", "pytest-cov", "ase"]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.0.0",
    "mkdocstrings[python]>=0.20.0",
    "mkdocs-glightbox>=0.3.0",
    "pymdown-extensions>=10.0.0",
    "mkdocs-git-revision-date-localized-plugin>=1.2.0",
    "mkdocs-awesome-pages-plugin>=2.8.0",
    "mkdocs-gen-files>=0.5.0",
    "griffe>=0.48.0",
]

# https://docs.deepmodeling.com/projects/deepmd/en/stable/development/create-a-model-pt.html
[project.entry-points."deepmd.pt"]
modifier_dipole_charge = "torch_dplr.dipole_charge:DipoleChargeModifier"

# [project.entry-points."deepmd"]
# modifier_dipole_charge_args = "torch_dplr.argcheck:modifier_dipole_charge_args"

[tool.setuptools_scm]
write_to = "torch_dplr/_version.py"
version_scheme = "release-branch-semver"
local_scheme = "node-and-date"

[tool.setuptools.packages.find]
include = ["torch_dplr*"]

[tool.setuptools.dynamic]
readme = { file = "README.md", content-type = "text/markdown" }

[tool.hatch.metadata.hooks.fancy-pypi-readme]
content-type = "text/markdown"

[[tool.hatch.metadata.hooks.fancy-pypi-readme.fragments]]
path = "README.md"
